Iteration on narrator MPC
