+++ backend/src/story/index.ts
@@ -1,2 +1,19 @@
 export { StoryService } from './storyService';
-export { MPCStoryAgents } from './mpcAgents'; 
+export { MPCStoryAgents } from './mpcAgents';
+
+// Export individual agents for direct use if needed
+export { StoryInitializer } from './storyInitializer';
+export { StoryOptimizer } from './storyOptimizer';
+export { StoryNarrator } from './storyNarrator';
+export { StoryPredictor } from './storyPredictor';
+
+// Export utility functions
+export { extractJsonFromResponse, createLoggerContext } from './mpcUtils';
+
+// Export interfaces for backward compatibility and direct use
+export type { 
+  StoryInitializerInput, 
+  StoryOptimizerInput, 
+  StoryNarratorInput, 
+  StoryPredictorInput 
+} from './mpcAgents';
